+++ src/alexey_khudoshin/motorradProject/Main.java
@@ -1,6 +1,6 @@
 package alexey_khudoshin.motorradProject;
 
-import motorradProject.model.Motorrad;
+import alexey_khudoshin.motorradProject.model.Motorrad;
 
 import java.util.*;
 import java.util.stream.Collectors;
