--- src/alexey_khudoshin/motorradProject/motorradProject/model/OneCylinder.java
@@ -1,19 +1,0 @@
-package alexey_khudoshin.motorradProject.motorradProject.model;
-
-import motorradProject.model.Motorrad;
-
-public class OneCylinder extends Motorrad {
-    String systemAddOiled;
-
-    public OneCylinder(String name, int power, String color, int engineVolume, int speed, String systemAddOiled) {
-        super(name, power, color, engineVolume, speed);
-        this.systemAddOiled = systemAddOiled;
-    }
-
-    @Override
-    public String toString() {
-        return "OneCylinder{" +
-                "systemAddOiled='" + systemAddOiled + '\'' +
-                '}';
-    }
-}
