--- src/alexey_khudoshin/motorradProject/motorradProject/model/TwoCylinder.java
@@ -1,17 +1,0 @@
-package alexey_khudoshin.motorradProject.motorradProject.model;//package DZ_27_10_2022;
-//
-//public class TwoCylinder extends Motorcycles{
-//    String systemDifOiled;
-//
-//    public TwoCylinder(String name, int power, String color, int engineVolume, int speed, String systemDifOiled) {
-//        super(name, power, color, engineVolume, speed);
-//        this.systemDifOiled = systemDifOiled;
-//    }
-//
-//    @Override
-//    public String toString() {
-//        return "TwoCylinder{" +
-//                "systemDifOiled='" + systemDifOiled + '\'' +
-//                '}';
-//    }
-//}
