--- src/alexey_khudoshin/motorradProject/motorradProject/model/TwoCylinderAirCooling.java
@@ -1,16 +1,0 @@
-package alexey_khudoshin.motorradProject.motorradProject.model;//package DZ_27_10_2022;
-//
-//public class TwoCylinderAirCooling extends TwoCylinder{
-//    String airCooling;
-//    public TwoCylinderAirCooling(String name, int power, String color, int engineVolume, int speed, String systemDifOiled, String airCooling) {
-//        super(name, power, color, engineVolume, speed, systemDifOiled);
-//        this.airCooling = airCooling;
-//    }
-//
-//    @Override
-//    public String toString() {
-//        return "TwoCylinderAirCooling{" +
-//                "airCooling='" + airCooling + '\'' +
-//                '}';
-//    }
-//}
